+++ packages/lexical-offset/src/index.ts
@@ -8,7 +8,6 @@
 
 import type {
   EditorState,
-  ElementNode,
   LexicalEditor,
   NodeKey,
   NodeMap,
@@ -17,6 +16,7 @@ import type {
 } from 'lexical';
 
 import {
+  $createChildrenArray as $createChildrenArray_,
   $createRangeSelection,
   $getNodeByKey,
   $isElementNode,
@@ -57,6 +57,7 @@ type OffsetInlineNode = {
 type OffsetNode = OffsetElementNode | OffsetTextNode | OffsetInlineNode;
 type OffsetMap = Map<NodeKey, OffsetNode>;
 
+/** @deprecated OffsetView has never worked correctly and will be removed */
 export class OffsetView {
   _offsetMap: OffsetMap;
   _firstNode: null | OffsetNode;
@@ -537,26 +538,10 @@ function $createOffsetChild(
   return firstNode;
 }
 
-export function $createChildrenArray(
-  element: ElementNode,
-  nodeMap: null | NodeMap,
-): Array<NodeKey> {
-  const children = [];
-  let nodeKey = element.__first;
-  while (nodeKey !== null) {
-    const node =
-      nodeMap === null ? $getNodeByKey(nodeKey) : nodeMap.get(nodeKey);
-    if (node === null || node === undefined) {
-      invariant(false, '$createChildrenArray: node does not exist in nodeMap');
-    }
-    children.push(nodeKey);
-    nodeKey = node.__next;
-  }
-  return children;
-}
-/** @deprecated renamed to {@link $createChildrenArray} by @lexical/eslint-plugin rules-of-lexical */
-export const createChildrenArray = $createChildrenArray;
+/** @deprecated moved to `lexical` */
+export const $createChildrenArray = $createChildrenArray_;
 
+/** @deprecated OffsetView has never worked correctly and will be removed */
 export function $createOffsetView(
   editor: LexicalEditor,
   blockOffsetSize = 1,
